implement test_scalar_compat for period
